# src/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Alias": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"alias": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"url",
					"alias"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Alias.url-or-alias_": {
				"properties": {
					"url": {
						"type": "string"
					},
					"alias": {
						"type": "string"
					}
				},
				"required": [
					"url",
					"alias"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CreateAliasResponse": {
				"$ref": "#/components/schemas/Pick_Alias.url-or-alias_"
			},
			"Pick_Alias.url_": {
				"properties": {
					"url": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CreateAliasBody": {
				"$ref": "#/components/schemas/Pick_Alias.url_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "url-shortener-api",
		"version": "0.1.0",
		"description": "URL shortener service",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Eunovo ",
			"email": "eunovo9@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/alias/{alias}": {
			"get": {
				"operationId": "GetUrk",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Alias"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alias",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/alias": {
			"post": {
				"operationId": "CreateAlias",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateAliasResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAliasBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}